Added Seeds for busniess & Customer sides

+++ BIGBACK/Domain/Order.cs
@@ -4,7 +4,7 @@
     {
         public DateTime OrderDateTime { get; set; }
         public string? OrderStatus { get; set; }
-        public int? OrderTotal { get; set; }
+        public decimal? OrderTotal { get; set; }
         public int CustomerId { get; set; }
     }
 }
